# JSON ingest performance