--- a/drivers/json/ingest_test.go
+++ b/drivers/json/ingest_test.go
@@ -20,6 +20,105 @@ import (
 	"github.com/neilotoole/sq/testh/tu"
 )
 
+func BenchmarkIngestJSONL_Flat(b *testing.B) {
+	// $ go test -count=10 -benchtime=5s -bench BenchmarkIngestJSONL_Flat > old.bench.txt
+	// # Make changes
+	// $ go test -count=10 -benchtime=5s -bench BenchmarkIngestJSONL_Flat > new.bench.txt
+	// $ benchstat old.bench.text new.bench.txt
+
+	// Either fpath (testdata file path) or input should be provided.
+	testCases := []struct {
+		name      string
+		fpath     string
+		input     string
+		wantRows  int
+		wantCols  []string
+		wantKinds []kind.Kind
+		wantErr   bool
+	}{
+		{
+			name:      "actor",
+			fpath:     "actor.jsonl",
+			wantRows:  sakila.TblActorCount,
+			wantCols:  sakila.TblActorCols(),
+			wantKinds: sakila.TblActorColKinds(),
+		},
+		{
+			name:      "film_actor",
+			fpath:     "film_actor.jsonl",
+			wantRows:  sakila.TblFilmActorCount,
+			wantCols:  sakila.TblFilmActorCols(),
+			wantKinds: sakila.TblFilmActorColKinds(),
+		},
+		{
+			name:      "actor_nested",
+			fpath:     "jsonl_actor_nested.jsonl",
+			wantRows:  4,
+			wantCols:  []string{"actor_id", "name_first_name", "name_last_name", "last_update"},
+			wantKinds: []kind.Kind{kind.Int, kind.Text, kind.Text, kind.Datetime},
+		},
+		{
+			name: "recs_medium",
+			input: `{"a": 1, "b": 1, "c": true, "d": "2020-06-11", "e": 2.0}
+{"a": 1.0, "b": 1, "c": false, "d": "2020-06-12", "e":2.01}`,
+			wantRows:  2,
+			wantCols:  []string{"a", "b", "c", "d", "e"},
+			wantKinds: []kind.Kind{kind.Int, kind.Int, kind.Bool, kind.Date, kind.Float},
+		},
+		{
+			name: "recs_small",
+			input: `{"b": 1}
+{"a": 1.1, "b": 2}`,
+			wantRows:  2,
+			wantCols:  []string{"b", "a"},
+			wantKinds: []kind.Kind{kind.Int, kind.Float},
+		},
+		{
+			name: "recs_null",
+			input: `{"a": null, "b": null}
+{"a": 1.1, "b": 2.0000}`,
+			wantRows:  2,
+			wantCols:  []string{"a", "b"},
+			wantKinds: []kind.Kind{kind.Float, kind.Int},
+		},
+	}
+
+	for _, tc := range testCases {
+		tc := tc
+
+		newRdrFn := func(_ context.Context) (io.ReadCloser, error) { //nolint:unparam
+			return io.NopCloser(strings.NewReader(tc.input)), nil
+		}
+
+		if tc.fpath != "" {
+			newRdrFn = func(_ context.Context) (io.ReadCloser, error) {
+				return os.Open(filepath.Join("testdata", tc.fpath))
+			}
+		}
+
+		b.Run(tc.name, func(b *testing.B) {
+			b.ReportAllocs()
+			b.ResetTimer()
+
+			for i := 0; i < b.N; i++ {
+				b.StopTimer()
+				th := testh.New(b, testh.OptNoLog())
+				src := th.Source(testsrc.EmptyDB)
+				grip := th.Open(src)
+
+				job := json.NewIngestJob(src, newRdrFn, grip, 0, true)
+
+				b.StartTimer()
+				err := json.IngestJSONL(th.Context, job)
+				if tc.wantErr {
+					require.Error(b, err)
+					return
+				}
+			}
+		})
+	}
+}
+
 func TestIngestJSONL_Flat(t *testing.T) {
 	t.Parallel()