Previously, JSON ingest jobs created, used, and closed a SQL prepared statement for _every single row_ of ingest data. Obviously this was wildly inefficient.

Now, the prepared statements are cached. There's still significant improvements that could be made, e.g. by using `driver.BatchInsert` (as seen with the CSV ingester), but this is a good start.

Benchmarks are better by `-23.76%` to `-77.62%` on the various metrics.

[wip.benchstat.txt](https://github.com/neilotoole/sq/files/14054113/wip.benchstat.txt)

```
goos: darwin
goarch: arm64
pkg: github.com/neilotoole/sq/drivers/json
                                 │ old.bench.txt │             new.bench.txt             │
                                 │    sec/op     │    sec/op     vs base                 │
IngestJSONL_Flat/actor-10           70.22m ± ∞ ¹   43.67m ± ∞ ¹        ~ (p=1.000 n=1) ²
IngestJSONL_Flat/film_actor-10       2.080 ± ∞ ¹    1.209 ± ∞ ¹        ~ (p=1.000 n=1) ²
IngestJSONL_Flat/actor_nested-10    1.907m ± ∞ ¹   1.463m ± ∞ ¹        ~ (p=1.000 n=1) ²
IngestJSONL_Flat/recs_medium-10     1.143m ± ∞ ¹   1.001m ± ∞ ¹        ~ (p=1.000 n=1) ²
IngestJSONL_Flat/recs_small-10      1.086m ± ∞ ¹   1.073m ± ∞ ¹        ~ (p=1.000 n=1) ²
IngestJSONL_Flat/recs_null-10      1165.9µ ± ∞ ¹   954.3µ ± ∞ ¹        ~ (p=1.000 n=1) ²
geomean                             8.595m         6.553m        -23.76%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

                                 │ old.bench.txt  │             new.bench.txt              │
                                 │      B/op      │     B/op       vs base                 │
IngestJSONL_Flat/actor-10          23.510Mi ± ∞ ¹   1.184Mi ± ∞ ¹        ~ (p=1.000 n=1) ²
IngestJSONL_Flat/film_actor-10     624.45Mi ± ∞ ¹   16.97Mi ± ∞ ¹        ~ (p=1.000 n=1) ²
IngestJSONL_Flat/actor_nested-10    505.9Ki ± ∞ ¹   161.6Ki ± ∞ ¹        ~ (p=1.000 n=1) ²
IngestJSONL_Flat/recs_medium-10     251.8Ki ± ∞ ¹   136.2Ki ± ∞ ¹        ~ (p=1.000 n=1) ²
IngestJSONL_Flat/recs_small-10      238.0Ki ± ∞ ¹   238.2Ki ± ∞ ¹        ~ (p=1.000 n=1) ²
IngestJSONL_Flat/recs_null-10       239.1Ki ± ∞ ¹   126.9Ki ± ∞ ¹        ~ (p=1.000 n=1) ²
geomean                             2.143Mi         491.1Ki        -77.62%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

                                 │ old.bench.txt │             new.bench.txt             │
                                 │   allocs/op   │  allocs/op    vs base                 │
IngestJSONL_Flat/actor-10           82.96k ± ∞ ¹   28.48k ± ∞ ¹        ~ (p=1.000 n=1) ²
IngestJSONL_Flat/film_actor-10     1617.1k ± ∞ ¹   323.9k ± ∞ ¹        ~ (p=1.000 n=1) ²
IngestJSONL_Flat/actor_nested-10    2.141k ± ∞ ¹   1.316k ± ∞ ¹        ~ (p=1.000 n=1) ²
IngestJSONL_Flat/recs_medium-10      953.0 ± ∞ ¹    661.0 ± ∞ ¹        ~ (p=1.000 n=1) ²
IngestJSONL_Flat/recs_small-10       535.0 ± ∞ ¹    536.0 ± ∞ ¹        ~ (p=1.000 n=1) ²
IngestJSONL_Flat/recs_null-10        574.0 ± ∞ ¹    389.0 ± ∞ ¹        ~ (p=1.000 n=1) ²
geomean                             6.619k         3.445k        -47.94%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05
```